service: Function for checking and updating service ordering
Update the service order checking with a lookup of preferred
technologies. If the new service is found to be preferred,
switch the order of the current default and the new service.

Do the same checks for the default service as currently in
online state default service checking; if the current default
service stays as default, return -EALREADY.

When neither service was found to be preferred, return -EAGAIN.