service: Changing default service was not correctly checked
Let the default_changed() function check whether the default service is
really changed as it is doing it correctly every time.

The extra check in the __connman_service_indicate_default() function is
causing problems if the user is moving services around. Then we might be
using the old DNS service data because DNS code did not get information
that default service was changed.

Fixes BMC#25027