dnsproxy: Delay cache removal
The idea here is to delay cache removal few seconds if there
are no cache users any more (refcount goes to 0). This is useful
for IPv6 RDNSS where new DNS servers are created right after old
one is removed. In this case we do not want to loose the cache
that still has perfectly valid data.