wispr: Add a reference counter mechanism on contexts
Basically, wispr provides a pointer (a wispr/portal context) to third parties
(agent, gweb...) which are answering asynchronously. However, if the technology
related to the service owning such context is disabled: this context is not
valid anymore, leading to a crash when third part answers after such action.

In order to avoid such issue, wispr implement an internal refcount system.

Fixes BMC#25479