session: Remove PeriodicConnect documentation
The idea was to let ConnMan know and optimize the connection
scheduling. Doing this per Session is a bit over-engineered.
The only use cases that have come up so far was for things like an
email client checking for new message.

Past experience showed that this kind of interface lead easily to
live locks. If there is need for such a feature, doing it on a per
technology would make more sense. For the time beeing it will be
removed. Later it could still be added to the Technology API.

If an application still wants to have a connection on regular
basis it can call the Session Connnect() method. ConnMan is still
able to defer any connection request.