session: Remove SessionMarker documentation
SO_MARK can only be set if the application has the CAP_NET_ADMIN
capability. We don't want to give that to all application for
good reasons and thus this approach won't work. Also touching all
applications to get this feature working is not a good idea. We
are going to use a cgroup controller to set SO_MARK and therefore
we don't need it in the Session API.