session: Move EmergencyCall to configuration plugin
Instead fallback using the configuration plugin for selecting
the emergency application.

Now we don't have to update all other sessions which makes
the whole code logic more straight forward. If a session is
marked as an emergency call and this application calls
Connect()/Disconnect() it will succedd and block all other
sessions when they try to call Connect() or Disconnect().