dnsproxy: Add reply host part length checking
Check that the first part of the name is not of zero length before
attempting to calculate the length of the domain part. Also ensure
the domain lenght checking does not run outside of the receive
buffer.

Also add debug messages for ids and lengths in order to pinpoint
any possible problems.