test: Renamed old VPN scripts
Renamed connect-vpn to connect-provider and
disconnect-vpn to remove-provider

In order to not confuse the user, change the name of the old
vpn connect/disconnect script. Now the name of the script
reflects the name of the manager dbus api.
The connect-provider will use the old ConnectProvider
manager dbus interface instead of directly connecting
to connman-vpnd.