dnsproxy: Use TCP connections only for the intended request
The DNS proxy code creates and uses a TCP connection only for a single
request at a time. A TCP DNS proxy is added to the end of the server
list, but should not be used for generic requests as the connection
will be closed down after the response to the initial request. Any
other requests sent over the connection will silently disappear.

Add a check to ignore TCP DNS servers for generic lookups.