core: Use interface index instead of interface name
Service and network code use only interface indexes. Convert the rest of
the code using interface names to use interface indexes instead. The files
affected are:
   include/resolver.h
   src/connman.h
   src/dnsproxy.c
   src/resolver.c
   src/rtnl.c
   src/service.c
   src/tethering.c

As the network and service code used interface indexes, the interface
name had to be looked up via a SIOCGIFNAME ioctl when needed. If a service
was removed due to the interface and network being taken down, the lookup
wouldn't work. This caused DNS servers to be left behind with broken sockets
since removing the DNS servers was done using the interface name.