service: Restart wispr on nameserver change
This is needed to make a service go online in the case where it was
already connected and then manual IPv4 & nameservers settings are
applied.  In that case, wispr is restarted with the new IP settings, but
the nameservers have not been set yet, so the wispr test fails and the
service remains in ready state.