session_policy_local: Create policy directory if necessary
Create the policy directory as first thing. If we don't do this
adding a watch on the non existing directory will fail and the
whole plugin is not loaded.

We need to figure out later how permissive the MODE of the directory
should be. Currently, we play safe and have it tied down.