ipconfig: Do not disable IPv6 in lower up state
Disabling IPv6 in lower up causes problems in IPv6 connectivity.
The interface IPv6 status can stay in disabled state because of
races. So this patch reverts the commit 4ce90440a70abce7de537777
and lets service IPv6 state go to READY when we get an auto
configured address for the interface.

Fixes BMC#25929