dnsproxy: Return proper return code if listener_table is missing
The listener_table can be NULL if we have started connman with -r
option. Because we return < 0 value to caller in this case,
the caller, e.g. the tethering code, has a possibility to change 
its functionality, e.g. the tethering code can fall back to some
other DNS servers.