network: Set service disconnect state correctly
If a connecting or associating network was disconnected, then
the corresponding service was left in associating state.
The fix is to check that we are not just in connected state,
but also connecting or associating network states are fine. In
this case the service state will be set correctly to idle.
The ip address etc. are only disabled/removed if we were in
connected state.

Fixes BMC#25981