netlink: Port ELL's version of netlink
It is time to have a common netlink code code for RTNL and also
for the upcoming NFACCT handling in the session core.

ELL provides a nice implementation for the netlink interface, so
let's port it over to use GLib and when ELL is ready to be used
we easily can start using it.