session: Store creation related data into a struct
During the creation period we need to keep track of data
which is only needed during this period. Let's rename user_config
to creation_data and store all the necessary data there. This includes
the session pointer so that we don't have to pass two void pointers
around in the callbacks which we will intrudoce soon.