dhcpv6: Initial IPv6 prefix delegation support
This patch adds start and stop prefix delegation functions.
The  __connman_dhcpv6_start_pd() is used to initiated the prefix
delegation. It will send the solicitation or rebind message to
the server depending on if we have old prefixes saved.
The __connman_dhcpv6_stop_pd() clears the prefix delegation data
and stops the delegation.
The normal DHCPv6 and prefix delegation can be going on at the
same time as they are used for different purposes. The PD is only
used if we are acting as a router i.e., when tethering.