session-test: Pass in struct test_fix into tests
Instead of using GSourceFunc signature test function signature
we struct test_fix directly. This is done by wrapping
the setup, test and teardown function inside session-utils.c.

That allows us also to get rid of the g_main_loop_run() dance during
setup and teardown. This is all done in session-utils.c.

In preperation to get rid of the gbooleans.