gdhcp: Fix byte order for send_discover()
Pass requested IP to send_discover() in host byte order.

DHCP options are byte-swapped on little endian systems when they're
added to the packet at transmission time, and inet_addr() gives us a
network byte ordered value, so we need to convert it to host byte
order before passing it to send_discover().