gdhcp: Report no lease found instead of starting IPv4LL
We need split the normal DHCP state machine from the IPv4LL one in
order to support DHCP retries after IPv4LL kicked in.

First thing to do is to report that we didn't find any lease in
normal DHCP mode when we reached the max retry counter (DISCOVER_RETRIES).

Second thing is if we reached the max counter we should not start the
IPv4LL state machine. That means dhcp.c needs to instantiate two clients,
one for DHCP and one for IPv4LL.