gsupplicant: Make sure it's possible to scan again when a scan is done
wpa_supplicant might send scan done signal, before uptading its state,
or the other way round. Due to that, it happends that on scan done,
interface state it still set as "scanning" though it's not anymore,
preventing the wifi plugin so raise a new scan right away after it gets
its scan callback called.

This fixes this issue. When scan is done, and if interface is still set
as scanning, we call scan finished callback, and we reset the scanning
state to false.