wifi: Postpone an hidden discovery if a scan is already running
This fixes a race condition when user tries to connect when connman is
proceeding with a scan already: at this point trying to connect to an
hidden AP will just fail due to a conflicting active scan that cannot be
ran. Thus, postoping such active scan, as soon as current scan is
finished, fixes it.

Reported by: Fabrice Peretti <fabrice.peretti@intel.com>