session: Use D-Bus owner to create unique paths for sessions
Two applications using '/foobar' as their notification path will both
end up with the same session path '/session/foobar'. When creating a
session the session path is used as the lookup key and an error is
returned if the session path already existed in the session hash
table. Fix this by using D-Bus owner information when creating the
unique session path.