main: Use gboolean type in GOptionEntry
The GOptionEntry options need to be of gboolean type as they are passed
as pointers. When the size of gboolean is different from the size of
bool, it will cause option variables to be overwritten by each other.

This was discovered by enabling the --nodaemon option which caused
option_dnsproxy to be overwritten to zero thereby disabling DNS
proxying altogether. Disabling DNS proxying forced ConnMan to write the
current nameserver entries directly into /etc/resolv.conf instead of
the loopback addresses. With DNS proxying disabled and no fallback
nameservers specified, no DNS entries were added to the DHCP server
messages and tethering functionality seemed to be broken also.