service: Use session information when autoconnecting
Collect the connected, connecting and pending checks into one place.
Continue with the next service in the list after that check of if a
service was connected. This ensures that all bearer types used by
sessions will be autoconnected. Since only bearer types are used,
only one service of the same type will be enabled. This is controlled
by the 'ignore' array. Finally, when the list of services reaches its
end, the code has to know whether something was being connected with
sessions active. This information is remembered in the
'autoconnecting' boolean.

As the preferred technologies can cover only a part of the available
technologies, an additional autoconnect using the whole service list
needs to be run one more time in run_auto_connect().