service: Do not clear 802.1x credentials when connecting
We cleared 802.1x credentials after starting the connect phase.
This is too early because connect is asynchronous and we might
remove the credentials before they are needed.

We need to clear the credentials if there is a connect error.
wpa_supplicant does not seem to return invalid key error for
WPA-EAP so we clear the credentials always if there is an error.

Reported-by: Thomas Green <TGreen2@Sorenson.com>