agent: Add multi agent support
Allow multiple agents to be registered to ConnMan. If an agent is
registered and a service is connected using the same DBus id (i.e.
from the same application), then that specific agent is consulted.

If the application has not registered an agent, another agent
selected by ConnMan is used to query the user. If no agent is
registered with ConnMan, then those operations that require user
input will fail.