dhcp: Store host and domain name in service struct
Store host and domain name when a lease becomes available in order to
use use them if the service is selected to be the default.

When creating a DHCP request use the value previously obtained. If such
a value does not exist, use the system host name.

For link-local autoconfiguration use the system host name just as
before.