session: Use a hash table to lookup sessions by service_entry
This is next step in cleaning up the struct service_entry.

Instead having a back link in struct service_entry to
service_entry use a global hash table which maps service_entries
to sessions.

This is an intermediate step to cleanup the session.c
file. Instead of doing everything in one big step we do baby steps.
That means also we need to keep stuff working in the steps. That's
the reason we add the global hash table and remove it later again.

Currently the only place we need this information is when we destroy
the entry.