session: Move service_entry to struct connman_session
Split the configuration of the session and the selected service.

The session configuration is tracked in struct session_info and
the selected service for given configuration is tracked in
struct service_entry. (Yes, struct service_entry will die, just
wait until we are there).

The struct service_entry is now directly a member of
struct connman_session.

While we are at let's get rid of some helper functions
and move the functionality (a one line) up to the caller function.