client: Fix unintended Manager monitor messages
Manager signals need to be captured for service and technology
completion. This has the unintended consequence that enabling any
other monitoring will also print out the Manager interface
messages. Fix this by providing a pointer to the relevant monitor
struct boolean in order to decide whether to forward the message
to the monitor command filter function or to stop its forwarding
in the command completion filter function.