device: Do not cleanup already provisioned interfaces
When devices are initialized, we check if the interface is
up and if it is, we cleanup the interface. This is not nice if
the interface is already provisioned and contains a valid IP
address.

If that is the case, then we check if IPv4 address and netmask
are the same as what we are trying to provision. If the addresses
are the same, then we do not init the interface because the it is
already setup. We only do this check for IPv4 addresses.